Made refactoring of parsing image code
